requirements.txt
#https://pypi.org/project/pyTelegramBotAPI/
pyTelegramBotAPI==4.18.1
python-decouple
